refactor: aboutpage to ts
