--- src/app/pages/about-page.utils/parse-url.js
@@ -1,5 +1,0 @@
-export const parseURL = function() {
-  let url = new URL(window.location);
-  let id = url.hash.split('/')[1];
-  return +id;
-}
